% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cricsheet_csv.R
\name{read_cricsheet_csv}
\alias{read_cricsheet_csv}
\title{Read a cricsheet CSV as a tidy tibble}
\usage{
read_cricsheet_csv(path)
}
\arguments{
\item{path}{Filepath to CSV}
}
\value{
a \code{tbl_df}
}
\description{
Read a cricsheet CSV as a tidy tibble
}
\details{
Will not recognise where "penalty runs" have been awarded, unless
they are included in extras.
}
